mods/ejs-esspm/cdk/requirements.txt
aws-cdk-lib==2.17.0
constructs>=10.0.0,<11.0.0
boto3==1.21.23
PyYAML==6.0
